requirements.txt
google-cloud-speech==2.12.0
# google-cloud-storage==2.1.0
python-telegram-bot==13.11
psycopg2-binary==2.9.3
